Separated ops and shifting Bit more const Comment tweaks
